sorty

+++ search_algorithms/0-linear.c
@@ -16,14 +16,14 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	int i;
+	size_t i; // Change the data type of i to size_t
 
 	if (array == NULL)
 		return (-1);
 
 	for (i = 0; i < size; i++)
 	{
-		printf("Value checked array[%lu] = %d\n", (unsigned long)i, array[i]);
+		printf("Value checked array[%lu] = %d\n", i, array[i]);
 		if (array[i] == value)
 			return (i);
 	}
